Y2038: improve printing of time settings
Avoid truncation errors when printing time settings.

Detected by Coverity. CID 1529622 (Use of 32-bit time_t)

+++ src/cache_cf.cc
@@ -2950,7 +2950,7 @@ parse_TokenOrQuotedString(char **var)
 static void
 dump_time_t(StoreEntry * entry, const char *name, time_t var)
 {
-    storeAppendPrintf(entry, "%s %d seconds\n", name, (int) var);
+    storeAppendPrintf(entry, "%s %ld seconds\n", name, static_cast<long>(var));
 }
 
 void
